fix(gym): split preview-param predicate so the server page can call it
CI e2e caught a bug missed locally: the gym landing (Server Component)
imported `isPreviewDemoActive` from `use-cardio-preview.ts`, which has
`'use client'` — Next 15+ refuses to call exports of a client module
from a Server Component. Local Vitest didn't trip because the unit
tests render hooks via `renderHook` (always client-side), and local
Playwright happened to hit a previously-cached build. CI's fresh dev
server hit the boundary check and returned 500 on every request,
which is why the e2e couldn't find the page's `<h1>`.

Fix: extract the predicate + URL-contract constants
(`isPreviewDemoActive`, `CARDIO_PREVIEW_PARAM`, `CARDIO_PREVIEW_VALUE`)
into a non-`'use client'` module at `lib/training-facility/preview-param.ts`.
The hook re-exports them for client-side ergonomics, but the server
page imports directly from the new module — bypassing the boundary.

A new `preview-param.test.ts` mirrors the import path the server uses,
so a future regression that re-routes the predicate through a client
module would surface in CI rather than only in browser smoke tests.
